+++ Editor/Utils/ViewDataKeyFoldoutStatus.cs
@@ -1,0 +1,9 @@
+namespace SaintsField.Editor.Utils
+{
+    public enum ViewDataKeyFoldoutStatus
+    {
+        None = 0,
+        Fold = 1,
+        Expand = 2,
+    }
+}