+++ Samples~/Scripts/AnimBehav/CustomAnimBehav.cs
@@ -1,0 +1,55 @@
+using SaintsField.Playa;
+using UnityEngine;
+
+namespace SaintsField.Samples.Scripts.AnimBehav
+{
+    [HelpURL("https://saintsfield.comes.today")]
+    public class CustomAnimBehav : StateMachineBehaviour
+    {
+        [LayoutStart("Layout?", ELayout.Horizontal)]
+        public int i1;
+        public int i2;
+
+        // OnStateEnter is called before OnStateEnter is called on any state inside this state machine
+        //override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+        //{
+        //
+        //}
+
+        // OnStateUpdate is called before OnStateUpdate is called on any state inside this state machine
+        //override public void OnStateUpdate(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+        //{
+        //
+        //}
+
+        // OnStateExit is called before OnStateExit is called on any state inside this state machine
+        //override public void OnStateExit(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+        //{
+        //
+        //}
+
+        // OnStateMove is called before OnStateMove is called on any state inside this state machine
+        //override public void OnStateMove(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+        //{
+        //
+        //}
+
+        // OnStateIK is called before OnStateIK is called on any state inside this state machine
+        //override public void OnStateIK(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+        //{
+        //
+        //}
+
+        // OnStateMachineEnter is called when entering a state machine via its Entry Node
+        //override public void OnStateMachineEnter(Animator animator, int stateMachinePathHash)
+        //{
+        //
+        //}
+
+        // OnStateMachineExit is called when exiting a state machine via its Exit Node
+        //override public void OnStateMachineExit(Animator animator, int stateMachinePathHash)
+        //{
+        //
+        //}
+    }
+}