+++ Samples~/Scripts/PlayableExposedReference/ExposedReferencePlayableTrack.cs
@@ -1,0 +1,9 @@
+using UnityEngine.Timeline;
+
+namespace SaintsField.Samples.Scripts.PlayableExposedReference
+{
+    [TrackClipType(typeof(ExposedReferencePlayableAsset))]
+    public class ExposedReferencePlayableTrack : TrackAsset
+    {
+    }
+}